Changes the implementation on checking the pending requests when closing the loadingModal
Now uses a factory variable to keep track of the pending requests that need the modal blocking the screen and closes it as soon as they are over. Doesn't leave the job of closing the modal to the last request, instead it's done by the last request that actually opened the modal.